FIX: 유저 고유 번호 자료형 변경으로 인한 로직 변경
카카오 로그인시 주어지는 카카오 유저 고유번호가 int로 받을 시 오버플로우 발생 -> Long으로 자료형 변경

+++ src/main/java/com/example/Taxi/domain/Member.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 public class Member {
     @Id @GeneratedValue
     private Long id;
-    private int identityNum;
+    private Long identityNum;
     private String nickname;
     private String accessTokenKaKao;
     private String refreshTokenKaKao;
@@ -23,7 +23,7 @@ public class Member {
     private Gender gender;
 
     @Builder
-    public Member(int identityNum, String nickname, String accessTokenKaKao, String refreshTokenKaKao, Gender gender) {
+    public Member(Long identityNum, String nickname, String accessTokenKaKao, String refreshTokenKaKao, Gender gender) {
         this.identityNum = identityNum;
         this.nickname = nickname;
         this.accessTokenKaKao = accessTokenKaKao;
